fix(analyzer): ensure consistent data types and rounding across all methods
- Add .round(2) to all decimal metrics (avg_clicks, median_clicks, std_clicks,
  median_position, std_position) in analyze_queries()
- Add .cast(pl.Int64) to all integer metrics (sum_clicks, sum_impressions, count)
- Apply consistent Int64 casting to comparison.py and recovery.py aggregations
- Update metrics.py summary methods to use proper types and rounding
- Ensure all DataFrame outputs have consistent Float64/Int64 types
- Update README.md changelog for v1.4.4

All decimal values now consistently use Float64 with 2 decimal places.
All integer values now consistently use Int64.

Closes data-type-consistency issue

+++ src/seo_analyzer/comparison.py
@@ -162,11 +162,11 @@ class ComparisonAnalyzer:
             })
         
         agg1 = df1.group_by(col_query).agg(
-            pl.col(metric).sum().alias(f'{metric}_p1')
+            pl.col(metric).sum().cast(pl.Int64).alias(f'{metric}_p1')
         )
         
         agg2 = df2.group_by(col_query).agg(
-            pl.col(metric).sum().alias(f'{metric}_p2')
+            pl.col(metric).sum().cast(pl.Int64).alias(f'{metric}_p2')
         )
         
         merged = agg1.join(agg2, on=col_query, how='full').fill_null(0)
